% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.r
\name{add_logo}
\alias{add_logo}
\title{Add logo}
\usage{
add_logo(plot, logo = NULL, size = 0.25)
}
\arguments{
\item{plot}{un graphique ggplot}

\item{logo}{un logo au format png, avec une transparence (si NULL utilise le logo OFCE)}

\item{size}{la taille du logo (relative)}
}
\value{
un graphique ggplot
}
\description{
Ajoute le logo de l'OFCE sur le graphique (inside donc)
}
\details{
Fin de vie. Ne plus utiliser
}
\examples{
\dontrun{
library(ggplot2)
plot <- ggplot(mtcars) +
        geom_point(aes(x=mpg, y=hp, size=cyl, col=gear)) +
        theme_ofce()
# plot |> add_logo()
}
}
