% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.r
\name{add_logo_ofce}
\alias{add_logo_ofce}
\title{Add logo}
\usage{
add_logo_ofce(plot, logo = ofce_logo, size = 0.04)
}
\arguments{
\item{plot}{un graphique ggplot}

\item{logo}{un logo au format png, avec une transparence}

\item{size}{la taille du logo (relative)}
}
\value{
un graphique ggplot
}
\description{
Ajoute le logo de l'OFCE en bas à droite du graphique
}
\examples{
\dontrun{
library(ggplot2)
plot <- ggplot(mtcars) +
    geom_point(aes(x=mpg, y=hp, size=cyl, col=gear)) +
    theme_ofce(base_family="Nunito")
# plot |> add_logo_ofce()
}
}
\seealso{
add_logo_ofce_inside
}
