% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.r, R/add_logo_inside.R
\name{add_logo_ofce_inside}
\alias{add_logo_ofce_inside}
\title{Add logo inside}
\usage{
add_logo_ofce_inside(plot, logo = ofce_logo, size = 0.25)

add_logo_ofce_inside(plot, logo = ofce_logo, size = 0.25)
}
\arguments{
\item{plot}{un graphique ggplot}

\item{logo}{un logo au format png, avec une transparence}

\item{size}{la taille du logo (relative)}
}
\value{
un graphique ggplot

un graphique ggplot
}
\description{
Ajoute le logo de l'OFCE sur le graphique (inside donc)

Ajoute le logo de l'OFCE sur le graphique (inside donc)
}
\examples{
\dontrun{
library(ggplot2)
plot <- ggplot(mtcars) +
        geom_point(aes(x=mpg, y=hp, size=cyl, col=gear)) +
        theme_ofce(base_family="Nunito")
# plot |> add_logo_ofce_inside()
}
\dontrun{
library(ggplot2)
plot <- ggplot(mtcars) +
        geom_point(aes(x=mpg, y=hp, size=cyl, col=gear)) +
        theme_ofce(base_family="Nunito")
# plot |> add_logo_ofce_inside()
}
}
\seealso{
add_logo_ofce

add_logo_ofce
}
