% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{add_OFCE_qext}
\alias{add_OFCE_qext}
\title{Extensions OFCE pour quarto}
\usage{
add_OFCE_qext()
}
\description{
Installe gentiment les extensions quarto suivantles templates OFCE
Trois templates sont prévus et s'insèrent dans un yml (ou dans la section yml) comme suit
format:
  ofce-html: default # on peut ajouter tous les champs html
  ofce-pdf: default # pour les pdf, à affiner
  ofce-revealjs: default # pour les présentations
}
