% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boards.R
\name{bd_read}
\alias{bd_read}
\title{Lit sur le board l'objet specifié}
\usage{
bd_read(obj)
}
\arguments{
\item{obj}{(chaine de caractère) nom de l'objet sur le board}
}
\value{
un objet R
}
\description{
Lit sur le board l'objet specifié
}
