% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boards.R
\name{bd_write}
\alias{bd_write}
\title{Ecrit sur le board}
\usage{
bd_write(
  obj,
  name = NULL,
  title = NULL,
  description = NULL,
  metadata = NULL,
  tags = NULL,
  versioned = NULL
)
}
\arguments{
\item{obj}{l'objet que l'on souhaite écrire}

\item{name}{le nom qu'il aura dans le pins::board, le nom de l'objet si ce n'est pas spécifié}

\item{title}{le titre du pins, optionel, mais aide à la compréhension de ce que l'objet contient}

\item{description}{une description, idem, pour faciliter l'usage}

\item{metadata}{des métadonnées utilisées par pins et stockées dans le board}

\item{tags}{idem, des tags}

\item{versioned}{verisoning du board (NULL par défaut, donc suit celui du board)}
}
\value{
(invisible) le nom du pins
}
\description{
Wrapper pour pins. la fonction suppose qu'il existe un blob Azure qui a été préalablement créé et
dont les éléments sont dans .Renviron (azure_url et azure_jeton) et dont les droits d'accès douvent être corrects.
}
