% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boards.R
\name{board}
\alias{board}
\title{Board avec Azure}
\usage{
board()
}
\value{
un board (de \code{{pins}})
}
\description{
Lit les dentifiants dans .Renviron et les utilise pour connecter avec Azure.
Les identifiants doivent être azure_url et un azure_jeton un jeton sas à obtenir sur le portail azure.
}
\details{
Cette fonction permet d'utiliser les fonctions de \code{{pins}}.
}
