% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.r
\name{date_jour}
\alias{date_jour}
\title{Date en jour}
\usage{
date_jour(
  date,
  locale = "fr_FR.UTF-8",
  tz = "Europe/Paris",
  short = FALSE,
  abbr = FALSE
)
}
\arguments{
\item{date}{la date}

\item{locale}{"fr_FR"}

\item{tz}{Time zone}

\item{short}{format court}

\item{abbr}{abrège les noms de mois}
}
\value{
une chaine de caractères
}
\description{
Transforme une date dans le format
2025-01-01 -> 1er janvier 2025
}
\examples{
date_jour("2025-10-01")
}
