% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.r
\name{date_jour_heure}
\alias{date_jour_heure}
\title{Date en jour / avec l'heure}
\usage{
date_jour_heure(date)
}
\arguments{
\item{date}{la date}
}
\value{
une chaine de caractères
}
\description{
Transforme une date dans le format trimestriel standard (revue de l'OFCE)
2025-01-01 -> 1 janvier 2025 12h02
}
\examples{
date_jour("2025-10-01")
}
