% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.r
\name{date_mois}
\alias{date_mois}
\title{Date en mois}
\usage{
date_mois(date)
}
\arguments{
\item{date}{la date}
}
\value{
une chaine de caractères
}
\description{
Transforme une date dans le format mensuel standard (revue de l'OFCE)
2025-01-01 -> janvier 2025
2025-04 -> avril 2025
}
\examples{
date_mois("2025-10-01")
}
