% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.r
\name{date_trim}
\alias{date_trim}
\title{Date en trimestriel}
\usage{
date_trim(date)
}
\arguments{
\item{date}{la date}
}
\value{
une chaine de caractères
}
\description{
Transforme une date dans le format trimestriel standard (revue de l'OFCE)
2025-01-01 -> T1 2025
2025-04 -> T2 2025
}
\examples{
date_trim("2025-10-01")
}
