% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{dernier_point}
\alias{dernier_point}
\title{fabricateur de dernier point pour les sources de graphiques}
\usage{
dernier_point(date, freq = "month", lang = "fr")
}
\arguments{
\item{date}{liste de date}

\item{freq}{soit day, month, quarter, year}

\item{lang}{langue de retour}
}
\value{
chaine de charactères
}
\description{
fabricateur de dernier point pour les sources de graphiques
}
