% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{f2si2}
\alias{f2si2}
\title{Utilise le SI pour formatter les nombres en fonction des mulitples de 10^3}
\usage{
f2si2(number, rounding = TRUE, digits = 1, unit = "median")
}
\arguments{
\item{number}{le nombre ou le vecteur à formatter (numérique)}

\item{rounding}{doit-on arrondir le chiffre ?}

\item{digits}{Combien de chiffres après la virgule}

\item{unit}{Arrondi soit à la "median" soit au "max"}
}
\value{
une chaine de caractère (character) formattée
}
\description{
Utilise le SI pour formatter les nombres en fonction des mulitples de 10^3
}
\examples{
f2si2(100000)

}
\seealso{
if2si2 (inverse la transformation)
}
