% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formate.R
\name{fmt_mds}
\alias{fmt_mds}
\title{Formate en milliards d'euros}
\usage{
fmt_mds(x, log = 3, digits = 1, suffix = " Mds €")
}
\arguments{
\item{x}{le nombre}

\item{log}{(defaut 3) différence en log10 entre l'origine et la destination 3 passe de millions à milliards}

\item{digits}{(defaut 1) nombre de chiffres après la virgule}

\item{suffix}{(defaut " Mds €") ajoute un suffix}
}
\value{
string
}
\description{
Semblable aux innombrables fonctions de formatage
formate en français ("," et " " comme séparateurs)
}
\examples{
fmt_mds(0.12)
}
