% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formate.R
\name{fmt_pct}
\alias{fmt_pct}
\title{Formate un nombre en pourcent}
\usage{
fmt_pct(x, digits = 0)
}
\arguments{
\item{x}{le nombre}

\item{digits}{(defaut 0) nombre de chiffres après la virgule}
}
\value{
string
}
\description{
Semblable aux innombrables fonctions de formatage
formate en français ("," et " " comme séparateurs)
}
\examples{
fmt_pct(0.12)
}
