% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formate.R
\name{fmt_val}
\alias{fmt_val}
\title{Formate un nombre}
\usage{
fmt_val(x, digits = 1, scale = 1, suffix = "")
}
\arguments{
\item{x}{le nombre}

\item{digits}{(defaut 1) nombre de chiffres après la virgule}

\item{scale}{(defaut 1) échelle (mulitplie par ce nombre)}

\item{suffix}{(defaut "") ajoute un suffix}
}
\value{
string
}
\description{
Semblable aux innombrables fonctions de formatage
formate en français ("," et " " comme séparateurs)
}
\examples{
fmt_val(10.56, 1, 1, " euros")
}
