% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2si.r
\name{if2si2}
\alias{if2si2}
\title{Transforme une chaîne formatté SI en nombre}
\usage{
if2si2(text)
}
\arguments{
\item{text}{le nombre ou le vecteur de nombres formattés à transformer}
}
\value{
un nombre
}
\description{
Transforme une chaîne formatté SI en nombre
}
\examples{
if2si2("100k")

}
