% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{init_qmd}
\alias{init_qmd}
\title{exécute le fichier rinit.R à la racine du projet}
\usage{
init_qmd(init = "rinit.r")
}
\arguments{
\item{init}{Nom du projet, "pres" par défaut,}
}
\description{
exécute le fichier rinit.R à la racine du projet
}
