% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafy.R
\name{load_graphe}
\alias{load_graphe}
\title{Lit un graphique sauvegardé}
\usage{
load_graphe(graphe)
}
\arguments{
\item{graphe}{(string) nom du graphique composé du dossier d'un tiret du nom du document d'un tiret et du label du graphique}
}
\value{
un ggplot
}
\description{
Lorsque \code{ofce.savegraph} ets non NULL, chaque graphique qui passe dans \code{girafy} est enregistré dans le dossier \code{ofce.savegraph}.
Son nom est formé en ajoutant le dossier qui le contient, le nom du fichier .qmd, le label du fichier (séparé par des tirets)
par exmple : "index-fig-psal" ou "france-synthese-fig-indicateurs"
}
