% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_graph.R
\name{load_gt}
\alias{load_gt}
\title{Lit un tableau sauvegardé}
\usage{
load_gt(object)
}
\arguments{
\item{object}{(string) nom du graphique composé du dossier d'un tiret du nom du document d'un tiret et du label du graphique}
}
\value{
un gt
}
\description{
Lorsque \code{ofce.savegraph} est non NULL, chaque tableau qui passe dans \code{save_gt} est enregistré dans le dossier \code{ofce.savegraph}.
Son nom est formé en ajoutant le dossier qui le contient, le nom du fichier .qmd, le label du fichier (séparé par des tirets)
par exmple : "index-fig-psal" ou "france-synthese-fig-indicateurs"
}
