% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.r
\name{logo_ofce}
\alias{logo_ofce}
\title{Logo de l'OFCE}
\usage{
logo_ofce(size = 1, background = "transparent")
}
\arguments{
\item{size}{(1 par défaut) la taille relative du logo (1 c'est bien)}

\item{background}{("transparent" par défaut) une couleur pour le fond}
}
\value{
un élément ggplot ( + )
}
\description{
Ajoute le logo de l'OFCE sur le graphique
}
\examples{
\dontrun{
library(ggplot2)
plot <- ggplot(mtcars) +
        geom_point(aes(x=mpg, y=hp, size=cyl, col=gear)) +
        theme_ofce() +
        logo_ofce()
}
}
