% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{make filename}
\usage{
make_filename(x, file = "", rep = "", env, ext)
}
\arguments{
\item{x}{nom de l'objet}

\item{file}{nom du fichier crée (par défaut nom de x)}

\item{rep}{stringr}

\item{env}{environnement}

\item{ext}{nom de l'extension du fichier (.png, .jpg,...)}
}
\value{
un nom de fichier
}
\description{
make filename
}
