% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_download.R
\name{margin_download}
\alias{margin_download}
\title{Bouton de téléchargement}
\usage{
margin_download(
  data,
  output_name = "donnees",
  label = "données",
  margin = TRUE,
  output_extension = getOption("ofce.output_extension"),
  prefix = getOption("ofce.output_prefix")
)
}
\arguments{
\item{data}{les données à téélcharger (un tibble donc)}

\item{output_name}{("données" par défaut) le nom du fichier de sortie}

\item{label}{("données") le nom du bouton qui apparaît dans le rendu du quarto}

\item{margin}{(TRUE) si FALSE le bouton est inline (non implémenté pour le moment)}

\item{prefix}{préfixe pour les fichiers téléchargés (\code{ofce.output_prefix} par défaut)}

\item{format}{format du ficher à télécharger, mis en option du package ( \code{ofce.output_extension} par défaut), peut être ".csv", ".xlsx"}
}
\value{
NULL (side effect : du markdown)
}
\description{
Cette fonction peremt d'ajouter un bouton de télchargement dans la marge d'un document quarto.
Elle repose sur \code{downloadthis} et prépare le div pour son affichage correct.
Elle doit être appelée dans un chunk \code{r} avec \code{results="asis"} impérativement.
}
