% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_download.R
\name{margin_download_file}
\alias{margin_download_file}
\title{Bouton de téléchargement d'un fichier}
\usage{
margin_download_file(
  data,
  output_name = fs::path_file(path),
  label = "données",
  margin = TRUE,
  prefix = getOption("ofce.output_prefix")
)
}
\arguments{
\item{output_name}{(\code{fs::path_file(path)} par défaut) le nom du fichier de sortie}

\item{label}{("données") le nom du bouton qui apparaît dans le rendu du quarto}

\item{margin}{(TRUE) si FALSE le bouton est inline (non implémenté pour le moment)}

\item{prefix}{préfixe pour les fichiers téléchargés (\code{ofce.output_prefix} par défaut)}

\item{path}{les données à téélcharger (un tibble donc)}
}
\value{
NULL (side effect : du markdown)
}
\description{
Cette fonction peremt d'ajouter un bouton de télchargement dans la marge d'un document quarto.
Elle repose sur \code{downloadthis::download_file} et prépare le div pour son affichage correct.
Elle doit être appelée dans un chunk \code{r} avec \code{results="asis"} impérativement.
}
