% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{minor_breaks_unity}
\alias{minor_breaks_unity}
\title{Breaks secondaires espacés d'une unité}
\usage{
minor_breaks_unity(limits)
}
\arguments{
\item{limits}{cette fonction reçoit les limites du ggplot en paramètre}
}
\value{
breaks
}
\description{
A utiliser dans scale_x_continuous
}
