% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{ofce_align_decimal}
\alias{ofce_align_decimal}
\title{Aligne sur le séparateur décimal (,)}
\usage{
ofce_align_decimal(data, columns = tidyselect::where(is.numeric))
}
\arguments{
\item{data}{le tableau gt passé en paramètre}

\item{columns}{les colonnes sélectionnées}
}
\value{
un tableau gt
}
\description{
Si aucune colonne n'est passée en paramètre, alors toutes les colonnes numériques sont alignées.
}
\examples{
head(mtcars) |> gt::gt() |> ofce_align_decimals(qsec)

}
\seealso{
\code{\link[=ofce_tab_options]{ofce_tab_options()}}, \code{\link[=ofce_caption]{ofce_caption()}}, \code{\link[=ofce_cols_fill]{ofce_cols_fill()}}, \code{\link[=ofce_spanners_bold]{ofce_spanners_bold()}}, \code{\link[=ofce_row_italic]{ofce_row_italic()}}, \code{\link[=ofce_fmt_decimal]{ofce_fmt_decimal()}}, \code{\link[=ofce_hide_col_pdf]{ofce_hide_col_pdf()}}
}
