% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{ofce_blog}
\alias{ofce_blog}
\title{installe un squelette de présentation}
\usage{
ofce_blog(dir = NULL, nom = NULL)
}
\arguments{
\item{dir}{Répertoire à créer "pres" par défaut}

\item{nom}{Nom du projet, "pres" par défaut,}
}
\description{
prépare un dossier, avec un exemple et les extensions nécessaires pour le formattage OFCE
}
