% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofce_caption.R
\name{ofce_caption}
\alias{ofce_caption}
\title{fabricateur de source}
\usage{
ofce_caption(object = NULL, ...)
}
\arguments{
\item{...}{autres paramètres}

\item{source}{texte de la source (sans le mot source qui est rajouté)}

\item{note}{texte de la note (sans le mot note qui est rajouté)}

\item{lecture}{texte de la note de lecture (sans le mot lecture qui est rajouté)}

\item{champ}{texte du champ (sans le mot champ qui est rajouté)}

\item{code}{texte du code (sans le mot code qui est rajouté)}

\item{xlab}{inclu le label de l'axe des x (pour le traduire avec marquee/glue)}

\item{ylab}{inclu le label de l'axe des y (pour le traduire avec marquee/glue)}

\item{subtitle}{inclu le label du sous titre (pour le traduire avec marquee/glue)}

\item{title}{inclu le titre (traduit aussi)}

\item{dpt}{dernier point connu}

\item{dptf}{fréquence du dernier point connu (day, month, quarter, year)}

\item{wrap}{largeur du texte en charactères (120 charactères par défaut, 0 ou NULL si on utilise marquee)}

\item{ofce}{(bool) si TRUE ajoute calculs OFCE à source, sinon rien, TRUE par défaut}

\item{author}{(bool) si TRUE ajoute calculs des auteurs à source, sinon rien, FALSE par défaut}

\item{srcplus}{(string) chaine (comme calculs OFCE) à ajouter à source (à la fin)}

\item{lang}{langue des textes (fr par défaut)}

\item{marquee_translate}{transforme ^x^ en {.sup x} et ~x~ en {.sub x}}

\item{glue}{applique glue avant toute chose}
}
\value{
un object de type variable
}
\description{
permet de construire un caption facilement avec un wrapping, le passage dans glue et la traduction marquee (pour ggplot)
}
\details{
on commence par la source, une note, et une lecture
}
\examples{
library(ggplot2)
ggplot(mtcars)+geom_point(aes(x=hp, y=disp)) + theme_ofce(marquee=TRUE) + ofce_caption(source="INSEE")
head(mtcars) |> gt::gt() |> ofce_caption(source="INSEE")
}
