% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofce_caption.R
\name{ofce_caption_gt}
\alias{ofce_caption_gt}
\title{fabricateur de source pour les tableaux}
\usage{
ofce_caption_gt(
  object,
  source = "",
  note = "",
  lecture = "",
  champ = "",
  code = "",
  dpt = "",
  subtitle = "",
  title = "",
  dptf = "month",
  wrap = ifelse(getOption("ofce.marquee"), 0, getOption("ofce.caption.wrap")),
  lang = getOption("ofce.caption.lang"),
  ofce = getOption("ofce.caption.ofce"),
  author = getOption("ofce.caption.author"),
  srcplus = getOption("ofce.caption.srcplus"),
  marquee_translate = ifelse(getOption("ofce.marquee"), TRUE,
    getOption("ofce.caption.marquee_translate")),
  glue = getOption("ofce.caption.glue"),
  ...
)
}
\arguments{
\item{object}{l'objet gt}

\item{source}{texte de la source (sans le mot source qui est rajouté)}

\item{note}{texte de la note (sans le mot note qui est rajouté)}

\item{lecture}{texte de la note de lecture (sans le mot lecture qui est rajouté)}

\item{champ}{texte du champ (sans le mot champ qui est rajouté)}

\item{code}{texte du code (sans le mot code qui est rajouté)}

\item{dpt}{dernier point connu}

\item{subtitle}{inclu le label du sous titre (pour le traduire avec marquee/glue)}

\item{title}{inclu le titre (traduit aussi)}

\item{dptf}{fréquence du dernier point connu (day, month, quarter, year)}

\item{wrap}{largeur du texte en charactères (120 charactères par défaut, 0 ou NULL si on utilise marquee)}

\item{lang}{langue des textes (fr par défaut)}

\item{ofce}{(bool) si TRUE ajoute calculs OFCE à source, sinon rien, TRUE par défaut}

\item{author}{(bool) si TRUE ajoute calculs des auteurs à source, sinon rien, FALSE par défaut}

\item{srcplus}{(string) chaine (comme calculs OFCE) à ajouter à source (à la fin)}

\item{marquee_translate}{transforme ^x^ en {.sup x} et ~x~ en {.sub x}}

\item{glue}{applique glue avant toute chose}

\item{...}{autres paramètres}
}
\value{
un objet gt (gt() |> ofce_caption_gt("INSEE"))
}
\description{
permet de construire un caption facilement avec un wrapping,
le passage dans glue
}
\details{
on commence par la source, une note, et une lecture
}
