% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{ofce_cols_fill}
\alias{ofce_cols_fill}
\title{Teinte les colonnes choisies en fond de couleur}
\usage{
ofce_cols_fill(data, columns, fill = getOption("ofce.tab.fill"))
}
\arguments{
\item{data}{le tableau gt passé en paramètre}

\item{columns}{les colonnes à teinter}

\item{fill}{la couleur (par défaut "#FFEDED")}
}
\value{
tableau gt
}
\description{
Teinte les colonnes choisies en fond de couleur
}
\examples{
head(mtcars) |> gt::gt() |> ofce_cols_fill(disp)
}
