% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{ofce_fmt_decimal}
\alias{ofce_fmt_decimal}
\title{Formate en français les nombres}
\usage{
ofce_fmt_decimal(
  data,
  columns = tidyselect::where(is.numeric),
  rows = tidyselect::everything(),
  decimals = 1
)
}
\arguments{
\item{data}{le tableau gt passé en paramètres}

\item{columns}{les colonnes sélectionnées (tout ce qui est numérique par défaut)}

\item{rows}{les lignes sélectionnées (tout par défaut)}

\item{decimals}{(le nombre de chiffres après la , 1 par défaut)}
}
\value{
un tableau gt
}
\description{
Formate en français les nombres
}
\examples{
head(mtcars) |> gt::gt() |> ofce_fmt_decimals()

}
\seealso{
\code{\link[=ofce_tab_options]{ofce_tab_options()}}, \code{\link[=ofce_caption]{ofce_caption()}}, \code{\link[=ofce_cols_fill]{ofce_cols_fill()}}, \code{\link[=ofce_spanners_bold]{ofce_spanners_bold()}}, \code{\link[=ofce_row_italic]{ofce_row_italic()}}, \code{\link[=ofce_align_decimal]{ofce_align_decimal()}}, \code{\link[=ofce_hide_col_pdf]{ofce_hide_col_pdf()}}
}
