% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{ofce_hide_col_pdf}
\alias{ofce_hide_col_pdf}
\title{Masque des colonnes en pdf}
\usage{
ofce_hide_col_pdf(tbl, col)
}
\arguments{
\item{tbl}{le gt passé en paramètre}

\item{col}{les colonnes sélectionnées}
}
\value{
un gt
}
\description{
Lorsqu'on utilise des colonnes qui peuvent bloquer en pdf, comme les nanoplots,
cette fonction sert à les enlever sélectivement.
}
\details{
Encore à tester avec typst
}
\examples{
head(mtcars) |> gt::gt() |> cols_hide_pdf(disp)

}
\seealso{
\code{\link[=ofce_tab_options]{ofce_tab_options()}}, \code{\link[=ofce_caption]{ofce_caption()}}, \code{\link[=ofce_cols_fill]{ofce_cols_fill()}}, \code{\link[=ofce_spanners_bold]{ofce_spanners_bold()}}, \code{\link[=ofce_row_italic]{ofce_row_italic()}}, \code{\link[=ofce_align_decimal]{ofce_align_decimal()}}, \code{\link[=ofce_fmt_decimal]{ofce_fmt_decimal()}}
}
