% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{ofce_palette}
\alias{ofce_palette}
\title{Palettes de couleurs pour la norme de la revue OFCE}
\usage{
ofce_palette(ncolor = 2)
}
\arguments{
\item{ncolor}{integer, nombre de couleurs}
}
\value{
une palette de couleurs (format HEX)
}
\description{
Palettes de couleurs pour la norme de la revue OFCE
}
