% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofce_palette_div.R
\name{ofce_palette_div}
\alias{ofce_palette_div}
\title{OFCE palette divergente}
\usage{
ofce_palette_div(colors = c("#005DA4", "#C51315"), n = NULL)
}
\arguments{
\item{colors}{un vecteur de deux couleurs à partir duquel la palette est générée.}

\item{n}{integer, nombre de couleurs}
}
\value{
un vecteur de couleurs de taille n
}
\description{
OFCE palette divergente
}
\examples{
ofce_palette_div(n = 4)

}
