% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofce_palette_new.R
\name{ofce_palette_new}
\alias{ofce_palette_new}
\title{Palette de cinq couleurs principales compatible avec les normes graphiques de la revue OFCE}
\usage{
ofce_palette_new(set = NULL)
}
\arguments{
\item{set}{choix du jeu de couleurs qui va être retenu}
}
\value{
un vector de couleurs de taille compris entre 2 et 5
}
\description{
Palette de cinq couleurs principales compatible avec les normes graphiques de la revue OFCE
}
\examples{
ofce_palette(n = 2)
}
