% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{ofce_palette_pays}
\alias{ofce_palette_pays}
\title{Palettes de couleur associées aux pays pour la norme de la revue OFCE}
\usage{
ofce_palette_pays(list_iso3 = list_iso.full)
}
\arguments{
\item{list_iso3}{chr, une liste de nom de pays}
}
\value{
une palette de couleur (format HEX)
}
\description{
Palettes de couleur associées aux pays pour la norme de la revue OFCE
}
