% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{ofce_palette_seq}
\alias{ofce_palette_seq}
\title{Palette de couleur associée de type sequentiel compatible avec les normes graphiques de la revue OFCE qui à partir de valeurs HEX produit une palette de n couleurs.}
\usage{
ofce_palette_seq(colors = c("#005DA4", "#C51315"), n = NULL)
}
\arguments{
\item{colors}{vecteur de deux couleurs à partir duquel la palette est générée}

\item{n}{integer, de couleurs}
}
\description{
Palette de couleur associée de type sequentiel compatible avec les normes graphiques de la revue OFCE qui à partir de valeurs HEX produit une palette de n couleurs.
}
