% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{ofce_row_bold}
\alias{ofce_row_bold}
\title{Met en gras une ligne}
\usage{
ofce_row_bold(data, row = everything())
}
\arguments{
\item{data}{le tableau gt passé en paramètre}

\item{row}{le sélecteur de la ligne}
}
\value{
un tableau gt
}
\description{
Met en gras une ligne
}
\examples{
head(mtcars) |> gt() |> ofce_row_bold(gear == 4)

}
\seealso{
\code{\link[=ofce_tab_options]{ofce_tab_options()}}, \code{\link[=ofce_caption]{ofce_caption()}}, \code{\link[=ofce_cols_fill]{ofce_cols_fill()}}, \code{\link[=ofce_spanners_bold]{ofce_spanners_bold()}}, \code{\link[=ofce_row_italic]{ofce_row_italic()}}, \code{\link[=ofce_align_decimal]{ofce_align_decimal()}}, \code{\link[=ofce_fmt_decimal]{ofce_fmt_decimal()}}, \code{\link[=ofce_hide_col_pdf]{ofce_hide_col_pdf()}}
}
