% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{ofce_row_italic}
\alias{ofce_row_italic}
\title{Mets une ligne en italique}
\usage{
ofce_row_italic(data, row = everything())
}
\arguments{
\item{data}{le tableau gt passé en paramètre}

\item{row}{le sélecteur de ligne}
}
\value{
un tableau gt
}
\description{
Mets une ligne en italique
}
\examples{
head(mtcars) |> gt::gt() |> ofce_row_italic(gear == 4)

}
\seealso{
\code{\link[=ofce_tab_options]{ofce_tab_options()}}, \code{\link[=ofce_caption]{ofce_caption()}}, \code{\link[=ofce_cols_fill]{ofce_cols_fill()}}, \code{\link[=ofce_spanners_bold]{ofce_spanners_bold()}}, \code{\link[=ofce_align_decimal]{ofce_align_decimal()}}, \code{\link[=ofce_fmt_decimal]{ofce_fmt_decimal()}}, \code{\link[=ofce_hide_col_pdf]{ofce_hide_col_pdf()}}
}
