% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{ofce_spanners_bold}
\alias{ofce_spanners_bold}
\title{Mets les spanners en gras}
\usage{
ofce_spanners_bold(data)
}
\arguments{
\item{data}{le tableau gt passé en paramètre}
}
\value{
un tableau gt
}
\description{
Mets les spanners en gras
}
\examples{
head(mtcars) |> gt::gt() |> gt::tab_spanner("spanner A", c(am, gear, carb)) |> ofce_spanners_bold()

}
\seealso{
\code{\link[=ofce_tab_options]{ofce_tab_options()}}, \code{\link[=ofce_caption]{ofce_caption()}}, \code{\link[=ofce_cols_fill]{ofce_cols_fill()}}, \code{\link[=ofce_row_italic]{ofce_row_italic()}}, \code{\link[=ofce_align_decimal]{ofce_align_decimal()}}, \code{\link[=ofce_fmt_decimal]{ofce_fmt_decimal()}}, \code{\link[=ofce_hide_col_pdf]{ofce_hide_col_pdf()}}
}
