% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\docType{data}
\name{palette.ofce}
\alias{palette.ofce}
\title{couleur pays de la charte graphique de l'OFCE}
\format{
Un dataframe avec 19 observations  and 2 variable:
\describe{
  \item{nom}{label, en français}
  \item{ISO}{le code iso, en 3 charactères}
  ...
}
}
\usage{
palette.ofce
}
\description{
Un dataset contenant les codes iso3 pour un ensemble de pays associés à un code couleur.
}
\keyword{datasets}
