% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_palette_degrades.R
\docType{data}
\name{palette_degrades}
\alias{palette_degrades}
\title{palette_degrades}
\format{
A data frame with 26 rows and 2 variables:
\describe{
\item{ Nom }{  character }
\item{ Code }{  character }
}
}
\source{
Source
}
\usage{
palette_degrades
}
\description{
Description.
}
\keyword{datasets}
