% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_palette_principale.R
\docType{data}
\name{palette_principale}
\alias{palette_principale}
\title{palette_principale}
\format{
A data frame with 8 rows and 4 variables:
\describe{
\item{ Nom }{  numeric }
\item{ Code }{  character }
\item{ R-V-B }{  character }
\item{ Observations }{  character }
}
}
\source{
Source
}
\usage{
palette_principale
}
\description{
Description.
}
\keyword{datasets}
