% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{quarto_wp}
\alias{quarto_wp}
\title{installe un squelette de document de travail}
\usage{
quarto_wp(nom = "wp", dir = nom)
}
\arguments{
\item{nom}{Nom du projet, "wp" par défaut,}

\item{dir}{Répertoire à créer "wp" par défaut}
}
\description{
prépare un dossier, avec un exemple et les extensions nécessaires pour le formattage OFCE
}
