% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_data.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{source_data}
\alias{source_data_refresh}
\alias{source_data_status}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sourcoise}{\code{\link[sourcoise:sourcoise]{source_data}}, \code{\link[sourcoise:sourcoise_refresh]{source_data_refresh}}, \code{\link[sourcoise:sourcoise_status]{source_data_status}}}
}}

