% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_graph.R
\name{save_gt}
\alias{save_gt}
\title{Enregistre un tableau}
\usage{
save_gt(object)
}
\arguments{
\item{object}{le ggplot}
}
\value{
le graphique, avec un effet de bord qui est le gt enregistré
}
\description{
L'enregistrement est fait dans le dossier spécifié dans les options
}
\seealso{
\code{\link[=save_object]{save_object()}}
}
