% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_graph.R
\name{save_object}
\alias{save_object}
\title{Enregistre un graphe}
\usage{
save_object(
  object,
  label = NULL,
  chunk = knitr::opts_current$get(),
  document = knitr::current_input(),
  id = NULL,
  dest = getOption("ofce.savegraph.dir"),
  ext = "ggplot"
)
}
\arguments{
\item{object}{le ggplot}

\item{label}{son label (mis à partir de knitr si possible)}

\item{chunk}{les infos de chunk (de knitr)}

\item{document}{le nom du document qmd (à partir de quarto)}

\item{id}{un id pour les tabset}

\item{dest}{le dossier de sauvegarde (ofce.savegraph)}
}
\value{
le graphique, avec un effet de bord qui est le ggplot enregistré
}
\description{
Enregistre un graphe
}
