% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_ofce_date.R
\name{scale_ofce_date}
\alias{scale_ofce_date}
\title{Echelle pour les x du format habituel des grahiques à date}
\usage{
scale_ofce_date(
  date_breaks = "5 years",
  labels = NULL,
  date_minor_breaks = "1 year",
  date_labels = "\%Y",
  guide = "minor_breaks",
  name = NULL,
  right_red = TRUE,
  ...
)
}
\arguments{
\item{date}{liste de date}

\item{freq}{soit day, month, quarter, year}

\item{lang}{langue de retour}
}
\value{
chaine de charactères
}
\description{
Echelle pour les x du format habituel des grahiques à date
}
