% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_data.R
\name{set_cache_rep}
\alias{set_cache_rep}
\title{répertoire de cache persistant}
\usage{
set_cache_rep(cache_rep = find_cache_rep())
}
\arguments{
\item{cache_rep}{(character) le répertoire}
}
\value{
rien
}
\description{
répertoire de cache persistant
}
\seealso{
Other source_data: 
\code{\link{clear_source_cache}()},
\code{\link{source_data}()},
\code{\link{source_data_refresh}()},
\code{\link{source_data_status}()}
}
\concept{source_data}
