% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{set_faicon_reveal}
\alias{set_faicon_reveal}
\title{Modifie l'icône de titre pour la présentation}
\usage{
set_faicon_reveal(path = ".", unicode = "f101")
}
\arguments{
\item{path}{chemin vers la présentation à modifier (le dossier qui contient la présentation)}

\item{icon}{code unicode hexa de l'icone (f101 par défaut)}
}
\description{
Attention cette fonction est un hack en attendant mieux
Elle est assez fragile et peut ne pas fonctionner
}
