% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{set_fontsize_reveal}
\alias{set_fontsize_reveal}
\title{Modifie la taille de police pour revealjs}
\usage{
set_fontsize_reveal(path = ".", size = 20)
}
\arguments{
\item{path}{chemin vers la présentation à modifier (le dossier qui contient la présentation)}

\item{size}{taille en "px" de la police pour revealjs (20 par défaut)}
}
\description{
Attention cette fonction est un hack en attendant mieux
Elle est assez fragile et peut ne pas fonctionner
}
