% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{set_justify}
\alias{set_justify}
\title{Modifie la justification des textes en HTML}
\usage{
set_justify(path = ".", justify = TRUE, ext = "ofce")
}
\arguments{
\item{path}{chemin vers le ou les documents HTML}

\item{justify}{(boolean) TRUE pour justifier, FALSE pour aligner à gauche}

\item{ext}{(string) extension des fichiers à modifier (ofce par défaut)}
}
\description{
Attention cette fonction est un hack en attendant mieux
Elle est assez fragile et peut ne pas fonctionner
}
