% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_pref.R
\name{setOFCERStudio}
\alias{setOFCERStudio}
\title{Reset RStudio prefs}
\usage{
setOFCERStudio()
}
\description{
Principalement, pas d'enregistrement de l'environnement, pipe natif, ragg en backend graphique
}
\examples{
if(interactive()) setOFCERStudio()

}
