% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{setup_blog}
\alias{setup_blog}
\title{installe un squelette pour le blog}
\usage{
setup_blog(dir = NULL, nom = NULL)
}
\arguments{
\item{dir}{Répertoire à créer "pres" par défaut}

\item{nom}{Nom du projet, "pres" par défaut,}
}
\description{
prépare un dossier, avec un exemple et les extensions nécessaires pour le formattage OFCE
}
