% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_pref.R
\name{setup_english}
\alias{setup_english}
\title{Set RStudio to english dictionary}
\usage{
setup_english()
}
\description{
Pour écrire en anglais dans RStudio
Attention un dictionnaire ne garantit en rien la qualité de l'expression
}
\examples{
if(interactive()) setup_english()

}
