% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_pref.R
\name{setup_french}
\alias{setup_french}
\title{Set RStudio to french dictionary}
\usage{
setup_french()
}
\description{
Pour écrire en français dans RStudio
Attention un dictionnaire ne garantit en rien la qualité de l'expression
}
\examples{
if(interactive()) setup_french()

}
