% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{setup_graph}
\alias{setup_graph}
\title{installe un squelette pour le graphique ofce}
\usage{
setup_graph(dir = NULL, nom = NULL)
}
\arguments{
\item{dir}{Répertoire à créer. Répertoire à créer. Par défaut, le qmd est placé dans le dossier de travail.}

\item{nom}{Nom du projet, "graph_ofce" par défaut,}
}
\description{
prépare un dossier, avec un exemple et les extensions nécessaires pour le formattage OFCE
}
