% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{setup_wp}
\alias{setup_wp}
\title{installe un squelette de document de travail}
\usage{
setup_wp(dir = NULL, nom = NULL)
}
\arguments{
\item{dir}{Répertoire à créer "wp" par défaut}

\item{nom}{Nom du projet, "wp" par défaut,}
}
\description{
prépare un dossier, avec un exemple et les extensions nécessaires pour le formattage OFCE
}
