% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_eurostat.R
\name{sna_clear_cache}
\alias{sna_clear_cache}
\title{Efface le cache sna}
\usage{
sna_clear_cache(cache = "./data/eurostat")
}
\arguments{
\item{cache}{le dossier du cache (par défaut /data/eurostat)}
}
\value{
rien, efface les fichiers
}
\description{
Efface le cache sna
}
\examples{
sna_clear_cache()
}
