% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_eurostat.R
\name{sna_eurostat}
\alias{sna_eurostat}
\title{Extrait les données eurostat SNA}
\usage{
sna_eurostat(
  dataset,
  ...,
  pivot = "auto",
  name = "values",
  cache = "./data/eurostat",
  select_time = NULL
)
}
\arguments{
\item{select_time}{}
}
\value{

}
\description{
Cette fonction permet d'extraire rapidement des données des comptes naitonaux sur eurostat.
Elle utilise le schéma d'accès \code{\link[eurostat:get_eurostat]{get_eurostat}}
}
