% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_eurostat.R
\name{sna_show}
\alias{sna_show}
\title{Infos sur une base sna Eurostat}
\usage{
sna_show(sna, lang = "fr", n = 100)
}
\arguments{
\item{sna}{le tibble téléchargé sur eurostat}
}
\value{
le tibble, invisible plus un effet de bord sur la console
}
\description{
Affiche les principales informations sur une base téléchargée sur Eurostat.
Les informations sont en partie stockées dans les attributs du tibble.
Ils peuvent être perdus en route.
}
\examples{
if(interactive()) {
data <- sna_get("nama_10_gdp")
sna_show(data)
}
}
