% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_data.R
\name{source_data_status}
\alias{source_data_status}
\title{Etat du cache de source_data}
\usage{
source_data_status(cache_rep = NULL, quiet = TRUE)
}
\arguments{
\item{data_rep}{le chemin vers le cache (défaut "_cache")}
}
\value{
tibble
}
\description{
Donne des informations sur le cache de source_data sous la forme d'un tibble
}
\seealso{
Other source_data: 
\code{\link{clear_source_cache}()},
\code{\link{set_cache_rep}()},
\code{\link{source_data}()},
\code{\link{source_data_refresh}()}
}
\concept{source_data}
