% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_ofce.r
\name{table_ofce}
\alias{table_ofce}
\title{Thème OFCE pour les tableaux pour {gt}}
\usage{
table_ofce(data, ...)
}
\arguments{
\item{data}{un objet gt}

\item{base_family}{character(1) string, police de charactère du thème (Stone sans par défaut)}
}
\value{
un objet gt, décoré
}
\description{
Applique le theme ofce compatible avec la norme de la Revue de l'OFCE
}
