% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabsetize.R
\name{tabsetize}
\alias{tabsetize}
\title{Tabsetize : fait un tabset à partir d'une liste}
\usage{
tabsetize(
  list,
  facety = TRUE,
  cap = TRUE,
  girafy = TRUE,
  asp = NULL,
  r = 1.5,
  o = 0.5,
  s = -1,
  pdf = getOption("ofce.tabsetize.pdf"),
  active = 1
)
}
\arguments{
\item{list}{liste des graphiques}

\item{facety}{(ne pas utiliser)}

\item{cap}{TRUE par défaut, insère une caption à la figure, spécifiée comme la caption du chunk (\verb{#| fig-cap: un titre}).}

\item{girafy}{TRUE par défaut, wrappe avec girafy (qui doit être défini donc dans le rinit.r)}

\item{asp}{aspect de ratio, mais privilégiez l'aspect ratio général (\verb{#| fig-asp: 1.1})}

\item{r}{rayon du cercle de hover pour girafy (paramètre \code{r} de girafy)}

\item{o}{opacité de hover_inv pour girafy (paramètre \code{o} de girafy)}

\item{s}{stroke with du hover pour girafy (paramètre \code{s} de girafy)}

\item{pdf}{si l'output est pdf, doit-on aficher tous les graphiques ("all", défaut) ou un seul ("one")}
}
\value{
string inserted in markdown
}
\description{
fabrique un tabset à partir d'une liste de graphiques (a priori des ggplots). Ne s'utilise que dans un qmd ou un markdown (ca n'a pas de sens sinon).
On peut également passer des chaines de caractères qui sont des chemins vers des images enregistrées. Cela peut permettre de faire un rendu plus vite.
Les noms de la liste sont utilisés pour les noms des onglets
Il est important de mettre l'option \code{results="asis"} au chunk (ou \verb{#| results: asis} dans le chunk)
Tabsetize prend en charge aussi les formats non intercatifs et affiche la liste dépliée (d'autres possibilités viendront dans le futur)
}
\details{
\code{r}, \code{s}, \code{o} peuvent être des listes qui s'apliqueront à chaque élément de la liste.
-1 pour chacune de ces valeurs implique pas de paramètre passé
}
