% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabsetize.R
\name{tabsetize2}
\alias{tabsetize2}
\title{Tabsetize2 : tabset à deux étages}
\usage{
tabsetize2(
  list,
  facety = TRUE,
  cap = TRUE,
  girafy = FALSE,
  asp = NULL,
  r = 1.5
)
}
\arguments{
\item{list}{liste des graphiques}

\item{facety}{(ne pas utiliser)}

\item{cap}{TRUE par défaut, insère une caption à la figure, spécifiée comme la caption du chunk (\verb{#| fig-cap: un titre}).}

\item{girafy}{FALSE par défaut, wrappe avec girafy (qui doit être défini donc dans le rinit.r)}

\item{asp}{aspect de ratio, mais privilégiez l'aspect ratio général (\verb{#| fig-asp: 1.1})}

\item{r}{rayon du cercle de hover pour girafy (paramètre \code{r} de girafy)}
}
\value{
string inserted in markdown
}
\description{
fabrique un tabset de tabset à partir d'une liste de liste de graphiques (a priori des ggplots, mais cela peut être des chemins).
Ne s'utilise que dans un qmd ou un markdown (ca n'a pas de sens sinon).
Les noms de la liste sont utilisés pour les noms des onglets.
Il est important de mettre l'option \code{results="asis"} au chunk (ou \verb{#| results: asis} dans le chunk)
Tabsetize prend en charge aussi les formats non intercatifs et affiche la liste dépliée (d'autres possibilités viendront dans le futur)
}
