% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_OFCE_blog.R
\name{template_OFCE_blog}
\alias{template_OFCE_blog}
\title{template_OFCE_blog}
\usage{
template_OFCE_blog(file_name = NULL)
}
\arguments{
\item{file_name}{Nom du nouveau fichier qmd et du sous-dossier dans lequel il est cr\\U+00E9\\U+00E9}
}
\value{
Un message de confirmation
}
\description{
charge un mod\\U+00E8le de document quarto pour un billet de blog
}
