% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_OFCE_presentation.R
\name{template_OFCE_presentation}
\alias{template_OFCE_presentation}
\title{template_OFCE_presentation}
\usage{
template_OFCE_presentation(file_name = NULL)
}
\arguments{
\item{file_name}{Nom du nouveau fichier qmd et du sous-dossier dans lequel il est cr\\U+00E9\\U+00E9}
}
\value{
Un message de confirmation
}
\description{
iconv(charge un mod\\U+00E8le de document quarto pour une presentation RevealJS, "UTF-8", "UTF-8")
}
