% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_OFCE_workingpaper.R
\name{template_OFCE_workingpaper}
\alias{template_OFCE_workingpaper}
\title{template_OFCE_workingpaper}
\usage{
template_OFCE_workingpaper(file_name = NULL)
}
\arguments{
\item{file_name}{Nom du nouveau fichier .qmd et du sous-dossier dans lequel il est cr\\U+00E9\\U+00E9}
}
\value{
Un message de confirmation
}
\description{
charge un mod\\U+00E8le pour un document de travail OFCE
}
