% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.gt_ofce.r
\name{theme.gt_ofce}
\alias{theme.gt_ofce}
\title{Thème OFCE pour les tableaux pour {gt}}
\usage{
theme.gt_ofce(data, largeur_colonne = NULL, ...)
}
\arguments{
\item{data}{un objet gt}

\item{largeur_colonne}{largeur de colonne (en px)}

\item{...}{autres arguments à passer dans la fonction}
}
\value{
un objet gt, décoré
}
\description{
Applique le theme ofce compatible avec la norme de la Revue de l'OFCE
}
