% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.gt_ofce2.R
\name{theme.gt_ofce_2}
\alias{theme.gt_ofce_2}
\title{Thème OFCE pour les tableaux pour {gt}}
\usage{
theme.gt_ofce_2(
  data,
  largeur_colonne = NULL,
  col_highlight = NULL,
  row_highlight = NULL,
  ...
)
}
\arguments{
\item{data}{un objet gt}

\item{largeur_colonne}{largeur de colonne (en px)}

\item{...}{autres arguments à passer dans la fonction}
}
\value{
un objet gt, décoré
}
\description{
Applique le theme ofce compatible avec la norme de la Revue de l'OFCE
}
