% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ofce_2.R
\name{theme_ofce.2}
\alias{theme_ofce.2}
\title{Thème OFCE 2}
\usage{
theme_ofce.2(
  base_size = getOption("ofce.base_size"),
  base_family = getOption("ofce.base_family"),
  ...
)
}
\arguments{
\item{base_size}{double(1) Taille des éléments texte du thème. Peut être donné globalement par options(ofce.base_size=12).}

\item{base_family}{character(1) string, police de charactère du thème (globalement et défaut options(ofce.base_family="Nunito"))}

\item{...}{paramètres passés à \code{\link[ggplot2]{theme}()}}
}
\value{
un thème qui peut être utilisé dans ggplot
}
\description{
Applique le theme ofce compatible avec la norme des working papers
d'autres éléments de la norme comme les couleurs, l'allure générale du graphique
sont à introduire par ailleurs
}
\seealso{
Other themes: 
\code{\link{theme_foundation}()},
\code{\link{theme_ofce}()},
\code{\link{theme_ofce_void}()}
}
\concept{themes}
