% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ofce.R
\name{theme_ofce}
\alias{theme_ofce}
\title{Thème OFCE
Applique le theme ofce compatible avec la norme de la Revue de l'OFCE
d'autres éléments de la norme comme les couleurs, l'allure générale du graphique
sont à introduire par ailleurs}
\usage{
theme_ofce(base_size = 12, base_family = "Stone sans")
}
\arguments{
\item{base_size}{double(1) Taille des éléments texte du thème}

\item{base_family}{character(1) string, police de charactère du thème (Stone sans par défaut)}
}
\value{
un thème qui peut être utilisé dans ggplot
}
\description{
Thème OFCE
Applique le theme ofce compatible avec la norme de la Revue de l'OFCE
d'autres éléments de la norme comme les couleurs, l'allure générale du graphique
sont à introduire par ailleurs
}
\examples{
library(ggplot2)
ggplot(mtcars) +
    geom_point(aes(x=mpg, y=hp, size=cyl, col=gear)) +
    theme_ofce(base_family="sans")
}
\seealso{
Other themes: 
\code{\link{theme_foundation}()}
}
\concept{themes}
