% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ofce_3.R
\name{theme_ofce_void}
\alias{theme_ofce_void}
\title{Thème OFCE void}
\usage{
theme_ofce_void(
  base_size = getOption("ofce.base_size"),
  base_family = getOption("ofce.base_family"),
  ...
)
}
\arguments{
\item{base_size}{double(1) Taille des éléments texte du thème. Peut être donné globalement par options(ofce.base_size=12).}

\item{base_family}{character(1) string, police de charactère du thème (globalement et défaut options(ofce.base_family="Nunito"))}

\item{...}{paramètres passés à \code{\link[ggplot2]{theme}()}}
}
\value{
un thème qui peut être utilisé dans ggplot
}
\description{
Applique le theme ofce compatible avec la norme de la Revue de l'OFCE
Convient pour des cartes par exemple
}
\seealso{
Other themes: 
\code{\link{theme_foundation}()},
\code{\link{theme_ofce}()},
\code{\link{theme_ofce.2}()}
}
\concept{themes}
