% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{uf2si2}
\alias{uf2si2}
\title{Formatte un vecteur en produisant des éléments distincts}
\usage{
uf2si2(number, rounding = TRUE, unit = "median", digits_max = 4)
}
\arguments{
\item{number}{le nombre ou le vecteur de nombre}

\item{rounding}{doit-on arrondir ?}

\item{unit}{Arrondi soit à la "median" soit au "max"}

\item{digits_max}{le nombre maximal de chiffres après la virgule}
}
\value{

}
\description{
Utilise f2si2 pour formatter un vecteur, et arrondi tant que les éléments formattés sont tous disctints.
Cela permet de les utiliser dans une échelle ou pour des noms.
Les nombres en entrée doivent être différents
}
\examples{
uf2si2(c(1000,1100,2000,2100))

}
