% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.r
\name{add_label_unit}
\alias{add_label_unit}
\title{Add label unit}
\usage{
add_label_unit(plot, ylabel = "")
}
\arguments{
\item{plot}{un graphique ggplot}

\item{ylabel}{string, un label}
}
\value{
un graphique ggplot
}
\description{
Ajoute un label pour les unités sur le plus grand des labels y
}
\examples{
\dontrun{
library(ggplot2)
plot <- ggplot(mtcars) +
        geom_point(aes(x=mpg, y=hp, size=cyl, col=gear)) +
        theme_ofce(base_family="Nunito")
# plot |> add_label_unit("horse power")
}
}
