% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate2.r
\name{annotate2}
\alias{annotate2}
\title{Variation de ggplot2::annotate pour intégrer une position}
\usage{
annotate2(
  geom,
  x = NULL,
  y = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  xend = NULL,
  yend = NULL,
  position = PositionIdentity,
  ...,
  na.rm = FALSE
)
}
\arguments{
\item{geom}{name of geom to use for annotation}

\item{x}{positioning aesthetics -
you must specify at least one of these.}

\item{y}{positioning aesthetics -
you must specify at least one of these.}

\item{xmin}{positioning aesthetics -
you must specify at least one of these.}

\item{xmax}{positioning aesthetics -
you must specify at least one of these.}

\item{ymin}{positioning aesthetics -
you must specify at least one of these.}

\item{ymax}{positioning aesthetics -
you must specify at least one of these.}

\item{xend}{positioning aesthetics -
you must specify at least one of these.}

\item{yend}{positioning aesthetics -
you must specify at least one of these.}

\item{position}{fonction position (comme position_nudge()) pour modifier la position de l'annotation}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
}
\value{

}
\description{
Variation de ggplot2::annotate pour intégrer une position
}
