% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafy.R
\name{girafy}
\alias{girafy}
\title{Rend un graphique interactif (avec ggiraph)}
\usage{
girafy(
  plot,
  r = 2.5,
  o = 0.5,
  out = getOption("ofce.savegraph"),
  id = NULL,
  ...
)
}
\arguments{
\item{plot}{le graphique ggplot}

\item{r}{(1.5) le rayon du zoom sur le point en cas de hover}

\item{o}{(0.5) l'opacité des autres éléments en cas de hover}

\item{out}{(NULL) répertoire pour enregistrer}

\item{id}{utilisé pour tabsetize}

\item{...}{autres options passées à \code{girafe_options()}}
}
\value{
un ggplot ou un objet ggiraph
}
\description{
Cette fonction prend en charge le passage par ggirapoh, en spécifiant les options importantes (dont les tooltips)
Possiblement, elle enregistre le graphique pour usage ultérieur (dans une présentation) ainf de garantir l'unicité des sources
Elle prend en charge également la sortie pdf si nécessaire
L'interactivité repose sur l'utilisation des fonctions de \code{ggiraph} \verb{geom_*_interactive()}
}
