% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{init_qmd}
\alias{init_qmd}
\title{exécute le fichier rinit.R à la racine du projet, ou dans _utils,
ou en dessous.
s'il ne le trouve pas il utilise une version par défaut,
stockée dans le package}
\usage{
init_qmd(init = "rinit.r", echo = FALSE, message = FALSE, warning = FALSE)
}
\arguments{
\item{init}{Nom du projet, "pres" par défaut,}
}
\description{
exécute le fichier rinit.R à la racine du projet, ou dans _utils,
ou en dessous.
s'il ne le trouve pas il utilise une version par défaut,
stockée dans le package
}
