% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{init_qmd}
\alias{init_qmd}
\title{exécute le fichier rinit.R à la racine du file, ou au dessus
s'il ne le trouve pas il utilise une version par défaut,
stockée dans le package}
\usage{
init_qmd(init = "rinit.r")
}
\arguments{
\item{init}{Nom du projet, "pres" par défaut,}
}
\description{
exécute le fichier rinit.R à la racine du file, ou au dessus
s'il ne le trouve pas il utilise une version par défaut,
stockée dans le package
}
