% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{ofce_caption}
\alias{ofce_caption}
\title{fabricateur de source pour les graphiques}
\usage{
ofce_caption(
  source = "Calculs OFCE",
  note = NULL,
  lecture = NULL,
  dpt = NULL,
  dptf = "month",
  wrap = 120,
  lang = "fr",
  ofce = TRUE
)
}
\arguments{
\item{source}{texte de la source (sans source)}

\item{note}{texte de la note}

\item{lecture}{texte de la note de lecture}

\item{dpt}{dernier point connu}

\item{dptf}{fréquence du dernier point connu (day, month, quarter, year)}

\item{wrap}{largeur du texte}

\item{lang}{langue des textes}

\item{ofce}{(bool) si TRUE ajoute calculs OFCE}
}
\value{
ggplot2 caption (ggplot() + ofce_caption())
}
\description{
permet de construire un caption facilement avec un wrapping
on commence par la source, une note, et une lecture
}
