% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofce_caption.R
\name{ofce_caption}
\alias{ofce_caption}
\title{fabricateur de source pour les graphiques}
\usage{
ofce_caption(
  source = NULL,
  note = NULL,
  lecture = NULL,
  dpt = NULL,
  dptf = "month",
  wrap = 110,
  lang = "fr",
  ofce = TRUE,
  author = NULL
)
}
\arguments{
\item{source}{texte de la source (sans le mot source qui est rajouté)}

\item{note}{texte de la note (sans le mot note qui est rajouté)}

\item{lecture}{texte de la note de lecture (sans le mot lecture qui est rajouté)}

\item{dpt}{dernier point connu}

\item{dptf}{fréquence du dernier point connu (day, month, quarter, year)}

\item{wrap}{largeur du texte en charactères (120 charactères par défaut)}

\item{lang}{langue des textes (fr par défaut)}

\item{ofce}{(bool) si TRUE ajoute calculs OFCE à source, sinon rien, TRUE par défaut}

\item{author}{(bool) si TRUE ajoute calculs des auteurs à source, sinon rien, FALSE par défaut}
}
\value{
ggplot2 caption (ggplot() + ofce_caption("INSEE"))
}
\description{
permet de construire un caption facilement avec un wrapping
on commence par la source, une note, et une lecture
}
