% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofce_palette_main.R
\name{ofce_palette}
\alias{ofce_palette}
\title{Palette de cinq couleurs principales compatible avec les normes graphiques de la revue OFCE}
\usage{
ofce_palette(n = 2)
}
\arguments{
\item{n}{number un nombre de couleurs à retenir pour la palette}
}
\value{
un vector de couleurs de taille compris entre 2 et 5
}
\description{
Palette de cinq couleurs principales compatible avec les normes graphiques de la revue OFCE
}
\examples{
ofce_palette(n = 3)
}
