% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofce_palette_div.R
\name{ofce_palette_div}
\alias{ofce_palette_div}
\title{Palette de couleurs associée de type divergente qui à partir de deux codes HEX génère une palette de n couleurs compatible avec les normes graphiques de la revue OFCE}
\usage{
ofce_palette_div(colors = c("#005DA4", "#C51315"), n = NULL)
}
\arguments{
\item{colors}{un vecteur de deux couleurs à partir duquel la palette est générée.}

\item{n}{integer, nombre de couleurs}
}
\value{
un vecteur de couleurs de taille n
}
\description{
Palette de couleurs associée de type divergente qui à partir de deux codes HEX génère une palette de n couleurs compatible avec les normes graphiques de la revue OFCE
}
\examples{
ofce_palette_div(n = 4)

}
