% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{ofce_palette_div}
\alias{ofce_palette_div}
\title{Palette de couleur associée de type sequentiel compatible avec les normes graphiques de la revue OFCE}
\usage{
ofce_palette_div(colors = c("#005DA4", "#C51315"), n = NULL)
}
\arguments{
\item{colors}{vecteur de deux couleurs à partir duquel la palette est générée}

\item{n}{integer, nombres de couleurs}
}
\value{
une palette de couleurs (format HEX)
}
\description{
Palette de couleur associée de type sequentiel compatible avec les normes graphiques de la revue OFCE
}
