% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofce_palette_main.R
\name{ofce_palette_main}
\alias{ofce_palette_main}
\title{OFCE palette principale}
\usage{
ofce_palette_main(n = 2)
}
\arguments{
\item{n}{number un nombre de couleurs à retenir pour la palette}
}
\value{
un vector de couleurs de taille compris entre 2 et 5
}
\description{
OFCE palette principale
}
\examples{
ofce_palette_main(n = 3)
}
