% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofce_palette_pays.R
\name{ofce_palette_pays}
\alias{ofce_palette_pays}
\title{Palette de couleur associée à un sous-ensemble de pays, compatible avec les normes graphiques de la revue OFCE}
\usage{
ofce_palette_pays(list_iso3 = palette.ofce$ISO)
}
\arguments{
\item{list_iso3}{chr, une liste de nom de pays}
}
\value{
une palette de couleur (format HEX)
}
\description{
Palette de couleur associée à un sous-ensemble de pays, compatible avec les normes graphiques de la revue OFCE
}
