% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_palette_pays.R
\docType{data}
\name{palette_pays}
\alias{palette_pays}
\title{palette_pays}
\format{
A data frame with 41 rows and 4 variables:
\describe{
\item{ ISO3 }{  character }
\item{ NOM }{  character }
\item{ Code }{  character }
\item{ R-V-B }{  character }
}
}
\source{
Source: Charte graphique OFCE (Juin 2024)
}
\usage{
palette_pays
}
\description{
Description.
}
\keyword{datasets}
