% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_data.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sourcoise}
\alias{sourcoise_refresh}
\alias{sourcoise_status}
\alias{sourcoise_clear}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sourcoise}{\code{\link[sourcoise]{sourcoise}}, \code{\link[sourcoise]{sourcoise_clear}}, \code{\link[sourcoise]{sourcoise_refresh}}, \code{\link[sourcoise]{sourcoise_status}}}
}}

