% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_graph.R
\name{save_ggplot}
\alias{save_ggplot}
\title{Enregistre un graphe}
\usage{
save_ggplot(object, id = NULL)
}
\arguments{
\item{object}{le ggplot}

\item{id}{utilisé par girafy et tabsetize}
}
\value{
le graphique, avec un effet de bord qui est le ggplot enregistré
}
\description{
L'enregistrement est fait dans le dossier spécifié dans les options
}
\seealso{
\code{\link[=save_object]{save_object()}}
}
