% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_pays.R
\name{scale_color_pays}
\alias{scale_color_pays}
\title{Palette de couleur associée à un sous-ensemble de pays, compatible avec les normes graphiques de la revue OFCE}
\usage{
scale_color_pays(list_iso3 = NULL, lang = NULL)
}
\arguments{
\item{list_iso3}{chr, une liste de nom de pays}
}
\value{
une palette de couleur (format HEX)
}
\description{
Palette de couleur associée à un sous-ensemble de pays, compatible avec les normes graphiques de la revue OFCE
}
