% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_pays.R
\name{scale_color_pays}
\alias{scale_color_pays}
\title{Palette de couleur associée à un sous-ensemble de pays, compatible avec les normes graphiques de la revue OFCE}
\usage{
scale_color_pays(
  format = NULL,
  lang = "fr",
  name = NULL,
  aesthetics = c("color", "fill"),
  ...
)
}
\arguments{
\item{format}{chr, le format des données ("iso3", "eurostat", "fr" pour les noms en clair, et tous les codes qu'accepte countrycode::countrycode)}

\item{lang}{chr, le choix de la langue c("fr","en")}

\item{name}{titre de la légende}

\item{aesthetics}{couleur ou fill ou les deux}

\item{...}{passé à scale_colour_manual}
}
\value{
un scale configuré
}
\description{
Palette de couleur associée à un sous-ensemble de pays, compatible avec les normes graphiques de la revue OFCE
}
