% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_pref.R
\name{setOFCERStudio}
\alias{setOFCERStudio}
\title{Reset RStudio prefs}
\usage{
setOFCERStudio(theme = "iPlastic", right_console = TRUE)
}
\arguments{
\item{theme}{Theme à utiliser (défaut "iPlastic")}

\item{right_console}{Met la console à droite (défaut TRUE)}
}
\description{
Principalement, pas d'enregistrement de l'environnement, pipe natif, ragg en backend graphique
}
\examples{
if(interactive()) setOFCERStudio()

}
