% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{setup_pres}
\alias{setup_pres}
\title{installe un squelette de présentation}
\usage{
setup_pres(dir = NULL, nom = NULL)
}
\arguments{
\item{dir}{Répertoire à créer. Par défaut, le qmd est placé dans le dossier de travail.}

\item{nom}{Nom du projet, "pres" par défaut,}
}
\description{
prépare un dossier, avec un exemple et les extensions nécessaires pour le formattage OFCE
}
