% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showmemoryuse.R
\name{showMemoryUse}
\alias{showMemoryUse}
\title{Utilisation mémoire par objet}
\usage{
showMemoryUse(
  sort = "size",
  decreasing = TRUE,
  limit = 10,
  envir = parent.frame()
)
}
\arguments{
\item{envir}{l'environement dans lequel sont listé les objets. Mieux vaut ne pas le toucher si on ne sait pas à quoi ça sert.}
}
\value{
Une liste des objets en mémoire, invisible et affiche dans la console cette liste
}
\description{
Permet de lister les plus gros objets en mémoire et de connaître leur empreinte.
}
\details{
copié de https://rdrr.io/github/zlfccnu/econophysics/ (merci!).
}
\examples{
showMemoryUse()

}
