% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna.R
\name{sna_check_cache}
\alias{sna_check_cache}
\title{MAJ le cache sna}
\usage{
sna_check_cache(cache = "./data/eurostat")
}
\arguments{
\item{cache}{le dossier du cache (par défaut /data/eurostat)}
}
\value{
la liste des bases mises \\U+E00E0 jour
}
\description{
MAJ le cache sna
}
\examples{
if(interactive()) sna_check_cache()
}
