% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_clear_cache.R
\name{sna_clear_cache}
\alias{sna_clear_cache}
\title{vide le cache sna}
\usage{
sna_clear_cache(cache = "./data/eurostat")
}
\arguments{
\item{cache}{le dossier du cache (par défaut /data/eurostat)}
}
\value{
rien
}
\description{
vide le cache sna
}
