% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_data.R
\name{source_data_refresh}
\alias{source_data_refresh}
\title{Exécute les sources sélectionnés}
\usage{
source_data_refresh(
  what = source_data_status(find_cache_rep()),
  cache_rep = find_cache_rep(),
  relative = getOption("ofce.source_data.relative"),
  force_exec = getOption("ofce.source_data.force_exec"),
  hash = getOption("ofce.source_data.hash")
)
}
\arguments{
\item{what}{un tibble issu de source_data (tout par défaut)}

\item{cache_rep}{le répertoire de cache}

\item{relative}{}

\item{force_exec}{(boléen) Si TRUE alors le code est exécuté ($FORCE_EXEC par défaut)}
}
\value{
un tibble de status
}
\description{
Exécute les sources sélectionnés
}
\seealso{
Other source_data: 
\code{\link{clear_source_cache}()},
\code{\link{set_cache_rep}()},
\code{\link{source_data}()},
\code{\link{source_data_status}()}
}
\concept{source_data}
