% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_data.R
\name{source_data_refresh}
\alias{source_data_refresh}
\title{Exécute les sources sélectionnés}
\usage{
source_data_refresh(
  what = NULL,
  cache_rep = NULL,
  force_exec = FALSE,
  hash = TRUE,
  unfreeze = TRUE,
  quiet = TRUE,
  init_qmd = TRUE,
  root = NULL
)
}
\arguments{
\item{what}{un tibble issu de source_data (tout par défaut)}

\item{cache_rep}{le répertoire de cache si il n'est pas évident}

\item{force_exec}{(boléen) Si \code{TRUE} alors le code est exécuté ($FORCE_EXEC par défaut)}

\item{hash}{(boléen) (\code{TRUE} par défaut) vérifie les hashs}

\item{unfreeze}{(boléen) (\code{TRUE} par défaut) essaye de unfreezé et uncaché les qmd dont les données ont été rafraichies}

\item{quiet}{reste silencieux}

\item{init_qmd}{(\code{TRUE} par défaut) exécute \code{ofce::init_qmd()}}

\item{root}{(\code{NULL} par défaut) essaye de trouver le root à partir du point d'exécution ou utilise celui fournit}
}
\value{
la liste des sources exécutés
}
\description{
Exécute les sources sélectionnés
}
\seealso{
Other source_data: 
\code{\link{clear_source_cache}()},
\code{\link{set_cache_rep}()},
\code{\link{source_data}()},
\code{\link{source_data_status}()}
}
\concept{source_data}
