% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_data.R
\name{source_data_status}
\alias{source_data_status}
\title{Etat du cache de source_data}
\usage{
source_data_status(data_rep = "_data")
}
\arguments{
\item{data_rep}{le chemin vers le cache (défaut "_cache")}
}
\value{
tibble
}
\description{
Donne des informations sur le cache de source_data sous la forme d'un tibble
}
