% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ofce.R
\name{theme_ofce}
\alias{theme_ofce}
\title{Thème OFCE}
\usage{
theme_ofce(
  base_size = getOption("ofce.base_size"),
  base_family = getOption("ofce.base_family"),
  ...
)

theme_ofce(
  base_size = getOption("ofce.base_size"),
  base_family = getOption("ofce.base_family"),
  ...
)
}
\arguments{
\item{base_size}{double(1) Taille des éléments texte du thème. Peut être donné globalement par options(ofce.base_size=12).}

\item{base_family}{character(1) string, police de charactère du thème (globalement et défaut options(ofce.base_family="Stone sans"))}

\item{...}{paramètres passés à \code{\link[ggplot2]{theme}()}}
}
\value{
un thème qui peut être utilisé dans ggplot

un thème qui peut être utilisé dans ggplot
}
\description{
Applique le theme ofce compatible avec la norme de la Revue de l'OFCE
d'autres éléments de la norme comme les couleurs, l'allure générale du graphique
sont à introduire par ailleurs

Applique le theme ofce compatible avec la norme de la Revue de l'OFCE
Convient pour des cartes par exemple
}
\examples{
library(ggplot2)
ggplot(mtcars) +
    geom_point(aes(x=mpg, y=hp, size=cyl, col=gear)) +
    theme_ofce(base_family="sans")
}
\seealso{
Other themes: 
\code{\link{theme_foundation}()}

Other themes: 
\code{\link{theme_foundation}()}
}
\concept{themes}
